Generic coin has been renamed to a skycoin
